valide json scheema & execption check & insert the new docs into db & refactore folder structure
