+++ README.md
@@ -1,2 +1,7 @@
 # angina-detection-using-afib
+
 ![cover_photo](./readmeimages/AFib-Animation.gif)
+
+```
+pip install -r requirements.txt
+```
