--- afib/services/service.py
@@ -1,9 +1,0 @@
-## /// buiness logic 
-
-## data processing
-
-##get data from databaase
-## return the predicted data
-
-class AtrialFibrillationServiceLayer:
-    pass
